更改处方明细表'总取药数量'字段数据类型

+++ jmp/mp-common/src/main/java/com/vpu/mp/common/pojo/shop/table/PrescriptionItemDo.java
@@ -42,7 +42,7 @@ public class PrescriptionItemDo implements Serializable {
     private Double    perTimeDosage;
     private String    perTimeDosageUnit;
     private Double    frequency;
-    private Integer   dragSumNum;
+    private Double    dragSumNum;
     private String    dragSumUnit;
     private String    goodsUseMemo;
     private String    goodsProductionEnterprise;
